requirements.txt
aiogram==3.0.0b5
aiohttp
pony
psycopg2-binary
python-dotenv
redis
starlette
sse_starlette
uvicorn
